phishingurldetector/requirements.txt
absl-py==2.1.0
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
django-environ==0.11.2
djangorestframework==3.15.1
emoji==2.2.0
fastapi==0.110.0
filelock==3.13.1
Flask==3.0.2
flatbuffers==23.5.26
folium==0.18.0
frozenlist==1.5.0
fsspec==2023.12.2
future==1.0.0
gevent==24.10.2
gitdb==4.0.11
GitPython==3.1.41
google==3.0.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.25.1
idna==3.6
imageio==2.33.1
ipykernel==6.29.0
ipython==8.21.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jupyter==1.0.0
jupyterlab==4.0.12
keras==2.15.0
lxml==5.1.0
Markdown==3.5.2
matplotlib==3.8.2
numpy==1.24.3
oauth2client==4.1.3
onxx==1.15.0
pandas==2.2.0
plotly==5.18.0
prometheus-client==0.19.0
protobuf==4.23.4
psycopg2==2.9.9
requests==2.31.0
scikit-learn==1.4.0
seaborn==0.13.2
sqlparse==0.4.4
tensorflow==2.15.0
torch==2.2.0
transformers==4.45.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
watchdog==3.0.0
